+++ reactivestate/src/commonMain/kotlin/com/ensody/reactivestate/CoroutineLauncher.kt
@@ -17,22 +17,19 @@ public interface CoroutineLauncher {
     /** The underlying [CoroutineScope] of this launcher. */
     public val launcherScope: CoroutineScope
 
-    /** Whether any loading state is currently doing something. */
-    public val isAnyLoading: LoadingStateTracker
-
     /**
      * The default loading tracker.
      *
      * Use [increment]/[decrement] to safely update the loading counter.
      */
-    public val generalLoading: MutableValueFlow<Int>
+    public val loading: MutableValueFlow<Int>
 
     /**
      * Launches a coroutine. Mark long-running coroutines by setting [withLoading] to loading state.
      *
      * @param context additional to [CoroutineScope.coroutineContext] context of the coroutine.
      * @param start coroutine start option. The default value is [CoroutineStart.DEFAULT].
-     * @param withLoading Tracks loading state for the (re-)computation. Defaults to [generalLoading].
+     * @param withLoading Tracks loading state for the (re-)computation. Defaults to [loading].
      *                    This should be `null` for long-running / never-terminating coroutines (e.g. `flow.collect`).
      * @param onError Optional custom error handler.
      * @param block the coroutine code which will be invoked in the context of the provided scope.
@@ -40,7 +37,7 @@ public interface CoroutineLauncher {
     public fun launch(
         context: CoroutineContext = EmptyCoroutineContext,
         start: CoroutineStart = CoroutineStart.DEFAULT,
-        withLoading: MutableValueFlow<Int>? = generalLoading,
+        withLoading: MutableValueFlow<Int>? = loading,
         onError: (suspend (Throwable) -> Unit)? = null,
         block: suspend CoroutineScope.() -> Unit
     ): Job